+++ R/package.R
@@ -4,6 +4,7 @@
 #'
 "_PACKAGE"
 
+#' @import checkmate
 #' @import ggfortify
 #' @import ggplot2
 #' @import ggplot2.utils