+++ NEWS.md
@@ -1,5 +1,8 @@
 # hermes 0.1.0.9000
 
+### New Features
+* Scatterplot of the gene expression values of two genes via `draw_scatterplot`.
+
 # hermes 0.1.0
 * First release of the `hermes` package, which contains classes, methods and functions to import, quality-check, filter, normalize, and analyze RNAseq counts data for differential expression. 
 * `hermes` is a successor of the `rnaseqTools` R package. The core functionality is built on the BioConductor ecosystem, especially the `SummarizedExperiment` class. New users should first begin by reading the "Introduction to `hermes`" vignette to become familiar with the `hermes` concepts.